+++ backend/app/schemas/price.py
@@ -1,12 +1,11 @@
 from pydantic import BaseModel
 from datetime import datetime
-from decimal import Decimal
 
 
 class PricePoint(BaseModel):
     timestamp: datetime
     bidding_zone: str
-    price_eur_mwh: Decimal | None
+    price_eur_mwh: float | None
     is_forecast: bool
 
     model_config = {"from_attributes": True}
